+++ iso20022-core/src/main/java/com/prowidesoftware/swift/model/mx/AppHdrParser.java
@@ -75,8 +75,8 @@ public class AppHdrParser {
 
     private static AppHdr parseHeaderFromSAXSource(
             final String xml, final String namespace, final MxReadParams params) {
-
-        SAXSource source = MxParseUtils.createFilteredSAXSource(xml, AppHdr.HEADER_LOCALNAME);
+        SAXSource source =
+                MxParseUtils.createFilteredSAXSource(MxParseUtils.makeXmlLenient(xml), AppHdr.HEADER_LOCALNAME);
 
         if (StringUtils.equals(LegacyAppHdr.NAMESPACE, namespace)) {
             // parse legacy AH